+++ lib/tasks/projects.rake
@@ -26,7 +26,7 @@ begin
     desc 'Clear locks from all projects'
     task :clear_lock => :environment do
     	require 'find'
-		Find.find(Rails.root.join('projects').to_s) { |path| FileUtils.rm Rails.root.join(path) if path =~ /\.lock.*/ }
+		Find.find(Rails.root.join('projects').to_s) { |path| FileUtils.rm Rails.root.join(path) if path =~ /\.lock.*/ } if Dir.exists? Rails.root.join('projects')
 	end
     desc 'Setup project assets'
     task :setup => :environment do
